requirements.txt
streamlit==1.32.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.202
torch==2.1.1+cpu
torchvision==0.16.1+cpu
